resources/code/workspace/langchain/langchain_agent_with_mcp/pyproject.toml
[project]
name = "langchain-agent-with-mcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastmcp==3.0.2",
    "httpx==0.28.1",
    "langchain==1.2.10",
    "langchain-mcp-adapters==0.2.1",
    "langchain-openai==1.1.10",
    "langgraph==1.0.9",
    "python-dotenv==1.2.1",
]
